+++ .rtx.toml
@@ -1,2 +1,3 @@
 [tools]
 ruby = "latest"
+node = "latest"
